+++ errors/errctx.go
@@ -33,11 +33,6 @@ func Newf(ctx context.Context, format string, args ...interface{}) error {
 	return Errorf(ctx, format, args...)
 }
 
-// Deprecated: Use `Wrap` or `Wrapf` instead of `Notef`. The library is able to unwrap mixed errors (wrapped with `errgo` or `github.com/pkg/errors`).
-func Notef(ctx context.Context, err error, format string, args ...interface{}) error {
-	return ErrCtx{ctx: ctx, err: errgo.Notef(err, format, args...)}
-}
-
 func Wrap(ctx context.Context, err error, message string) error {
 	return ErrCtx{ctx: ctx, err: errors.Wrap(err, message)}
 }
@@ -55,10 +50,6 @@ func Errorf(ctx context.Context, format string, args ...interface{}) error {
 func RootCtxOrFallback(ctx context.Context, err error) context.Context {
 	var lastCtx context.Context
 
-	type causer interface {
-		Cause() error
-	}
-
 	// Unwrap each error to get the deepest context
 	for err != nil {
 		// First check if the err is type of `*errgo.Err` to be able to call `Underlying()`
@@ -71,18 +62,21 @@ func RootCtxOrFallback(ctx context.Context, err error) context.Context {
 			continue
 		}
 
-		cause, ok := err.(causer)
+		cause, ok := err.(interface{ Cause() error })
 		if ok {
 			err = cause.Cause()
 			continue
 		}
 
-		// if err is type of `ErrCtx` unwrap it by getting errCtx.err
-		ctxerr, ok := err.(ErrCtx)
+		// If err is matching the `ErrCtx` interface type, unwrap it and get its context.
+		// We compare with the interface to match different versions of ErrCtx package.
+		ctxerr, ok := err.(interface {
+			Ctx() context.Context
+			Unwrap() error
+		})
 		if ok {
-			err = ctxerr.err
+			err = ctxerr.Unwrap()
 			lastCtx = ctxerr.Ctx()
-
 			continue
 		}