+++ crypto/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## To be Released
 
+## v1.1.1
+
+* chore(go): corrective bump - Go version regression from 1.24.3 to 1.24
+
+## v1.1.0
+
+* chore(go): upgrade to Go 1.24
+
 ## v1.0.0
 
 * Initial release of the lib