+++ fs/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## To be Released
 
+## v1.1.1
+
+* chore(go): corrective bump - Go version regression from 1.24.3 to 1.24
+
+## v1.1.0
+
+* chore(go): upgrade to Go 1.24
 * bump github.com/spf13/afero from 1.8.2 to 1.9.3
 
 ## v1.0.2