+++ gomock_generator/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## To be Released
 
+## v1.4.1
+
+* chore(go): corrective bump - Go version regression from 1.24.3 to 1.24
+
+## v1.4.0
+
+* chore(go): upgrade to Go 1.24
 * build(deps): bump github.com/Scalingo/go-utils/logger from 1.1.1 to 1.2.0
 * build(deps): bump github.com/sirupsen/logrus from 1.8.1 to 1.9.0
 * build(deps): bump github.com/urfave/cli from 1.22.9 to 1.22.10