+++ influx/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## To be Released
 
+## v1.3.1
+
+* chore(go): corrective bump - Go version regression from 1.24.3 to 1.24
+
+## v1.3.0
+
+* chore(go): upgrade to Go 1.24
 * build(deps): bump github.com/influxdata/influxdb from 1.9.7 to 1.11.0
 * build(deps): bump github.com/stretchr/testify from 1.8.0 to 1.8.1