+++ nsqproducer/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## To be Released
 
+## v1.3.1
+
+* chore(go): corrective bump - Go version regression from 1.24.3 to 1.24
+
+## v1.3.0
+
+* chore(go): upgrade to Go 1.24
+
+## v1.2.0
+
 * build(deps): bump github.com/Scalingo/go-utils/logger from 1.1.1 to 1.2.0
 * build(deps): bump github.com/sirupsen/logrus from 1.8.1 to 1.9.0
 * build(deps): bump github.com/gofrs/uuid from 4.2.0+incompatible to 4.3.1+incompatible