+++ storage/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## To be Released
 
+## v1.7.2
+
+* build(deps): rollback github.com/aws/aws-sdk-go-v2 to v1.32.8
+
+## v1.7.1
+
+* chore(go): corrective bump - Go version regression from 1.24.3 to 1.24
+
+## v1.7.0
+
+* chore(go): upgrade to Go 1.24
+
 ## v1.6.0
 
 * feat(s3): Pass the retry options to the AWS SDK on top of our custom retrier.